refactoring searchable restore
